requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic-settings==2.7.1
discord.py==2.4.0
python-dateutil==2.9.0.post0
